requirements.txt
django-tables2==1.21.2
